+++ scripts/run_model.py
@@ -1,17 +1,15 @@
-import json
 import os
 import sys
 
 import fire
-import pandas as pd
 
 print(sys.path)
 print("Current working directory:", os.getcwd())
 sys.path.append(os.getcwd())
 
-from src.prompting.messages import build_messages
 from src.simulation.inference import simulate_whole_survey
 from src.simulation.models import load_opinion_gpt, load_llama, change_adapter, change_persona
+from src.simulation.utils import huggingface_login, load_survey, save_survey
 
 
 LOAD_MODEL = {
@@ -23,18 +21,6 @@ CHANGE_SUBGROUP = {
     "llama": change_persona,
 }
 
-
-def huggingface_login() -> None:
-    from dotenv import load_dotenv
-    from huggingface_hub import login
-
-    load_dotenv()
-    HF_TOKEN = os.environ.get("HF_TOKEN")
-    if HF_TOKEN:
-        login(token=HF_TOKEN)
-        print("Successfully logged in to Hugging Face!")
-    else:
-        print("Token is not set. Please save a token in the .env file.")
 
 def main(model_name: str, directory: str, subgroup: str, filename: str = "variables.csv", device: str = "cuda:2"):
 
@@ -50,19 +36,6 @@ def main(model_name: str, directory: str, subgroup: str, filename: str = "variab
         "respondents": respondents,
     }
     save_survey(survey_run, directory)
-
-
-def load_survey(directory: str, file_name: str) -> dict[str, str]:
-    survey_df = pd.read_csv(os.path.join(directory, file_name))
-    survey = build_messages(survey_df)
-    print("Successfully loaded survey!")
-    return survey
-
-
-def save_survey(simulated_survey: dict[str, dict], directory: str):
-    with open(os.path.join(directory, "survey_run.json"), "w") as f:
-        json.dump(simulated_survey, f)
-        print("Successfully saved simulated responses!")
 
 
 if __name__ == "__main__":
